+++ packages/apptive_grid_form/lib/src/translation/l10n/translation_de.dart
@@ -88,6 +88,8 @@ class ApptiveGridLocalizedTranslation extends ApptiveGridTranslation {
   String get addressGeocodingFailed =>
       "Position konnte nicht automatisch ermittelt werden";
   @override
+  String get recordVideo => "Video aufnehmen";
+  @override
   String resourceGroupLabel(DataResourceMetaType metaType) =>
       switch (metaType) {
         DataResourceMetaType.grid => "Grids und Ansichten",